requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg[binary]>=3.2
pillow>=10.0
django-cors-headers>=4.4
python-dotenv>=1.0
requests